Validação de dados CRUD Pacientes

+++ src/controllers/pacientesController.js
@@ -4,7 +4,7 @@ const pacientesController = {
   async listarPacientes(req, res){
     try {
       const listaPacientes = await Pacientes.findAll();
-      res.status(200).json(listaPacientes);
+      res.status(201).json(listaPacientes);
     } 
     catch (error) {
       res.json(`Falha ao listar pacientes!`);
@@ -20,7 +20,7 @@ const pacientesController = {
       if(!paciente){
         res.status(404).json("Paciente não encontrado");
       }
-      res.status(200).json(paciente);
+      res.status(201).json(paciente);
     } 
     catch (error) {
       res.json(`Falha ao listar pacientes!`);
@@ -31,10 +31,16 @@ const pacientesController = {
   async cadastrarPacientes(req, res){
     try {
       const { nome, email, idade } = req.body;
+
+      const existePaciente = await Pacientes.count({where: { email }});
+      if(existePaciente){
+        return res.status(400).json("email já cadastrada!");
+      }
+
       const novoPaciente = await Pacientes.create({
         nome, email, idade
       });
-      res.status(200).json(novoPaciente);
+      res.status(201).json(novoPaciente);
     } 
     catch (error) {
       res.json(`Falha ao cadastrar paciente!`);
@@ -45,11 +51,18 @@ const pacientesController = {
   async excluirPacientes(req, res){
     try {
       const { id } = req.params;
+
+      const existeId = await Pacientes.count({where: { id }});
+
+      if(!existeId){
+        return res.status(400).json("Paciente não encontrado!");
+      }
+
       await Pacientes.destroy({
         where: { id }
       });
 
-      res.status(200).json("Paciente removido!");
+      res.status(201).json("Paciente removido!");
     } catch (error) {
       res.json("Falha ao remover paciente!")
       console.error(error);
@@ -61,6 +74,12 @@ const pacientesController = {
     const { id } = req.params;
     const { nome, email, idade } = req.body;
 
+    const existeId = await Pacientes.count({where: { id }});
+
+    if(!existeId){
+      return res.status(400).json("Paciente não encontrado!");
+    }
+
     const pacienteAtualizado = await Pacientes.update(
     { nome, email, idade },
     {
@@ -69,7 +88,7 @@ const pacientesController = {
       }
     }
     );
-     res.status(200).json("Paciente atualizado com sucesso!");
+     res.status(201).json("Paciente atualizado com sucesso!");
     } 
     catch (error) {
       res.status(404).json("Verfique os dados e tente novamente");
